move to ping.go
